+++ tests/testthat/test_statfitools.R
@@ -28,7 +28,7 @@ context("get_class")
 
 test_that("get_class functions return data.frame",{
   expect_is(sf_get_class_key("kuntaryhmitys", 2014), "data.frame")
-  expect_is(sf_get_reg_key("maakunta", 2014), "data.frame")
+  expect_is(sf_get_reg_keytable("Maakunta"), "data.frame")
   expect_is(sf_get_class("maakunta", 2013), "data.frame")
 })